services/location-service/requirements.txt
Django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
python-dotenv>=1.0
psycopg[binary]>=3.2
redis>=5.0
django-prometheus>=2.3
